+++ src/test/java/com/github/tommyettinger/ds/test/AllProductIntVectorHashTest.java
@@ -204,9 +204,9 @@ public class AllProductIntVectorHashTest {
 					}
 
 					@Override
-					public void setHashMultiplier (long hashMultiplier) {
-						this.hashMultiplier = hashMultiplier | 1L;
-						hm = (int)this.hashMultiplier;
+					public void setHashMultiplier (int hashMultiplier) {
+						this.hashMultiplier = hashMultiplier | 1;
+						hm = this.hashMultiplier;
 						resize(keyTable.length);
 					}
 				};
